requirements.txt
# Core dependencies
python-dotenv>=0.19.0
PyGithub>=1.55
aiohttp>=3.8.0
pyyaml>=6.0.1

# Trading
pybit>=2.4.0
pandas>=1.5.0
numpy>=1.23.0
ccxt>=2.0.0

# Machine Learning
scikit-learn>=1.0.2
torch>=1.12.0
transformers>=4.21.0

# Automation & Calendar
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2
selenium>=4.5.0
beautifulsoup4>=4.11.1

# Development
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.982

# Monitoring & Logging
prometheus-client>=0.14.1
grafana-api>=1.0.3
elasticsearch>=8.4.0
